chore: Moves profile to commerce and adds missing cookies
