+++ packages/api/src/platforms/vtex/clients/commerce/types/Profile.ts
@@ -15,5 +15,5 @@ export interface ProfileAddress {
   complement: string
   reference: string
   receiverName: string
-  geoCoordinate: [number][number]
+  geoCoordinate: [number, number]
 }
